added clients

+++ src/Controller/Controller.java
@@ -32,7 +32,7 @@ import java.util.logging.Logger;
  */
 public class Controller {
 
-	boolean enable_web=true;
+	boolean enable_web = true;
 	Configuration _config;
 	Slot[] _slots;
 	Host[] _hosts;
@@ -48,18 +48,20 @@ public class Controller {
 	GenericScheduler generic_scheduler;
 	WebRequestStatsSlot[][][] _webRequestSlotStats;
 
+	int[][][][] running_allocations; // [n][p][v][s]
 
 	int[][] _webRequestPattern; // [p][s]: p=provider, s=service
 
-	int hosts_number; 
-	int providers_number; 
-	int vm_types_number; 
-	int services_number; 
+	int hosts_number;
+	int providers_number;
+	int vm_types_number;
+	int services_number;
 	int slots_number;
 
 	Timer stats_timer;
 
-	int[][][] total_requests; 
+	int[][][] total_requests;
+	int slot = 0;
 
 	Controller(Configuration config, Host[] hosts, Slot[] slots, Provider[] _provider) {
 
@@ -75,8 +77,8 @@ public class Controller {
 		this._webUtilities = new WebUtilities(config);
 		this._cplexData = new SchedulerData(config);
 		this._providers = _provider;
-
-		this.generic_scheduler=new GenericScheduler(config);
+		this.running_allocations = new int[hosts_number][providers_number][vm_types_number][services_number];
+		this.generic_scheduler = new GenericScheduler(config, this);
 		initializeController();
 		_cplexData.initializeWebRequestMatrix(_webRequestPattern);
 
@@ -84,13 +86,14 @@ public class Controller {
 
 	private void initializeController() {
 
-		this._webRequestSlotStats = new WebRequestStatsSlot[_config.getSlotsNumber()][providers_number][services_number];
+		this._webRequestSlotStats = new WebRequestStatsSlot[_config
+				.getSlotsNumber()][providers_number][services_number];
 		this._webRequestPattern = new int[providers_number][services_number];
 
 		for (int slot = 0; slot < slots_number; slot++) {
 			for (int p = 0; p < this.providers_number; p++) {
 				for (int s = 0; s < this.services_number; s++) {
-					_webRequestSlotStats[slot][p][s] = new WebRequestStatsSlot();
+					_webRequestSlotStats[slot][p][s] = new WebRequestStatsSlot(slot, p, s);
 				}
 			}
 		}
@@ -100,11 +103,11 @@ public class Controller {
 				_webRequestPattern[p][s] = (int) _config.getArrivals_generator()[p][s].get("estimated_requests");
 			}
 		}
-		total_requests= new int[providers_number][vm_types_number][services_number];
+		total_requests = new int[providers_number][vm_types_number][services_number];
 		for (int p = 0; p < this.providers_number; p++) {
 			for (int v = 0; v < this.vm_types_number; v++) {
 				for (int s = 0; s < this.services_number; s++) {
-					total_requests[p][v][s]=0;
+					total_requests[p][v][s] = 0;
 				}
 			}
 		}
@@ -113,13 +116,14 @@ public class Controller {
 	@SuppressWarnings("unused")
 	void Run(int slot) throws IOException {
 
+		this.slot = slot;
 		System.out.println("******* Slot:" + slot + " Controller Run *******");
 		updateServiceRequestPattern(slot);
 
 		scheduler = new Scheduler(_config);
 
-		//		if (false)
-		//			startNodesStatsTimer(slot); // for Statistics updates
+		// if (false)
+		// startNodesStatsTimer(slot); // for Statistics updates
 
 		try {
 
@@ -130,16 +134,18 @@ public class Controller {
 				}
 			}
 			int[][][] vmRequestMatrix = loadVMRequestMatrix(slot);
-			Utilities.updateRequestStats2Db(slot,_config,vmRequestMatrix,total_requests);
+			Utilities.updateRequestStats2Db(slot, _config, vmRequestMatrix, total_requests);
 
-			System.out.println("REQUESTS Matrix:"+Arrays.deepToString(vmRequestMatrix));
+			System.out.println("REQUESTS Matrix:" + Arrays.deepToString(vmRequestMatrix));
 
 			// Load VM Deactivation Matrix
-			int[][][][] vms2DeleteMatrix=prepareVmDeleteMatrix(slot);
-			System.out.println("DELETE Matrix:"+Arrays.deepToString(vms2DeleteMatrix));
+			int[][][][] vms2DeleteMatrix = prepareVmDeleteMatrix(slot);
+			System.out.println("DELETE Matrix:" + Arrays.deepToString(vms2DeleteMatrix));
 			reduceRunningAllocation(vms2DeleteMatrix);
-			destroyServices(slot);
-			Thread.sleep(10000);
+			if (!_config.getSimulation_mode()) {
+				destroyServices(slot);
+				Thread.sleep(10000);
+			}
 
 			// Update WebRequest Pattern
 			// int[][] requestPattern=Utilities.findRequestPattern(_config);
@@ -160,24 +166,22 @@ public class Controller {
 				else
 					System.out.print("No scheduling algorithm is defined");
 
-				Utilities.updateActivationStats(slot,_config,activationMatrix);
-
+				Utilities.updateActivationStats(slot, _config, activationMatrix);
 
 			}
 
-			System.out.println("ACTIVATION Matrix:"+Arrays.deepToString(activationMatrix));
-
+			System.out.println("ACTIVATION Matrix:" + Arrays.deepToString(activationMatrix));
 
 			scheduler.updateData(_cplexData, activationMatrix);
 			CplexResponse cplexResponse = updatePenaltyAndUtility(_cplexData, activationMatrix);
 
 			// ----------- Update Statistics Object
 			double net_benefit = cplexResponse.getNetBenefit();
-			System.out.println("NET_BENEFIT: "+net_benefit);
+			System.out.println("NET_BENEFIT: " + net_benefit);
 
-			// ----------- Create VMs Actual or Objects)
-
-			createAllServices(slot,activationMatrix);
+			// ----------- Create VMs Actual)
+			if (!_config.getSimulation_mode()) 
+				createAllServices(slot, activationMatrix);
 
 		} catch (Exception ex) {
 			Logger.getLogger(Controller.class.getName()).log(Level.SEVERE, null, ex);
@@ -239,47 +243,57 @@ public class Controller {
 		}
 	}
 
-	//	class StatisticsTimer extends TimerTask {
+	// class StatisticsTimer extends TimerTask {
 	//
-	//		int slot;
-	//		int numberOfMachineStatsPerSlot = _config.getNumberOfMachineStatsPerSlot();
+	// int slot;
+	// int numberOfMachineStatsPerSlot =
+	// _config.getNumberOfMachineStatsPerSlot();
 	//
-	//		StatisticsTimer(int slot) {
-	//			this.slot = slot;
-	//		}
+	// StatisticsTimer(int slot) {
+	// this.slot = slot;
+	// }
 	//
-	//		public void run() {
+	// public void run() {
 	//
-	//			if (_currentInstance < numberOfMachineStatsPerSlot) {
+	// if (_currentInstance < numberOfMachineStatsPerSlot) {
 	//
-	//				//		Utilities.updateSimulatorStatistics(_config, slot, _currentInstance);
+	// // Utilities.updateSimulatorStatistics(_config, slot, _currentInstance);
 	//
-	//				_currentInstance++;
+	// _currentInstance++;
 	//
-	//			} else {
-	//				stats_timer.cancel();
-	//			}
+	// } else {
+	// stats_timer.cancel();
+	// }
 	//
-	//		}
+	// }
 	//
-	//	}
+	// }
 
-	//	private void startNodesStatsTimer(int slot) {
+	// private void startNodesStatsTimer(int slot) {
 	//
-	//		int statsUpdateInterval = _config.getSlotDuration() / _config.getNumberOfMachineStatsPerSlot();
+	// int statsUpdateInterval = _config.getSlotDuration() /
+	// _config.getNumberOfMachineStatsPerSlot();
 	//
-	//		stats_timer = new Timer();
-	//		_currentInstance = 0;
+	// stats_timer = new Timer();
+	// _currentInstance = 0;
 	//
-	//		if (_config.getSlotDurationMetric().equals(ESlotDurationMetric.milliseconds.toString()))
-	//			stats_timer.scheduleAtFixedRate(new StatisticsTimer(slot), 0, statsUpdateInterval);
-	//		else if (_config.getSlotDurationMetric().equals(ESlotDurationMetric.seconds.toString()))
-	//			stats_timer.scheduleAtFixedRate(new StatisticsTimer(slot), 0, statsUpdateInterval * 1000);
-	//		else if (_config.getSlotDurationMetric().equals(ESlotDurationMetric.minutes.toString()))
-	//			stats_timer.scheduleAtFixedRate(new StatisticsTimer(slot), 0, 60 * statsUpdateInterval * 1000);
-	//		else if (_config.getSlotDurationMetric().equals(ESlotDurationMetric.hours.toString()))
-	//			stats_timer.scheduleAtFixedRate(new StatisticsTimer(slot), 0, 3600 * statsUpdateInterval * 1000);
-	//	}
+	// if
+	// (_config.getSlotDurationMetric().equals(ESlotDurationMetric.milliseconds.toString()))
+	// stats_timer.scheduleAtFixedRate(new StatisticsTimer(slot), 0,
+	// statsUpdateInterval);
+	// else if
+	// (_config.getSlotDurationMetric().equals(ESlotDurationMetric.seconds.toString()))
+	// stats_timer.scheduleAtFixedRate(new StatisticsTimer(slot), 0,
+	// statsUpdateInterval * 1000);
+	// else if
+	// (_config.getSlotDurationMetric().equals(ESlotDurationMetric.minutes.toString()))
+	// stats_timer.scheduleAtFixedRate(new StatisticsTimer(slot), 0, 60 *
+	// statsUpdateInterval * 1000);
+	// else if
+	// (_config.getSlotDurationMetric().equals(ESlotDurationMetric.hours.toString()))
+	// stats_timer.scheduleAtFixedRate(new StatisticsTimer(slot), 0, 3600 *
+	// statsUpdateInterval * 1000);
+	// }
 
 	private void addVmRequestsPerService(ServiceRequest _serviceRequest, int slot) {
 
@@ -298,18 +312,17 @@ public class Controller {
 			_serviceRequest.getVms_requested()[v] = _vms[v];
 		}
 
-
 	}
 
 	@SuppressWarnings("unused")
-	private void createAllServices(int slot,int[][][][] activation_matrix) {
+	private void createAllServices(int slot, int[][][][] activation_matrix) {
 		LoadService load_service_object;
 		Thread thread;
 		int vms_number = 0;
 
-		if(enable_web){
+		if (enable_web) {
 			try {
-				load_service_object = new LoadService(_config,slot,activation_matrix);
+				load_service_object = new LoadService(_config, slot, activation_matrix);
 				thread = new Thread(load_service_object);
 				thread.start();
 				Thread.sleep(0);
@@ -324,16 +337,13 @@ public class Controller {
 	@SuppressWarnings("unused")
 	private void destroyServices(int slot) throws InterruptedException {
 
-
-		if(enable_web){
+		if (enable_web) {
 			DestroyService deleter = new DestroyService(_slots[slot]);
 			Thread thread = new Thread(deleter);
 			thread.start();
 			Thread.sleep(5);
 		}
 	}
-
-
 
 	private CplexResponse updatePenaltyAndUtility(SchedulerData data, int[][][][] activationMatrix) {
 
@@ -365,7 +375,6 @@ public class Controller {
 
 		System.out.println();
 
-
 		CplexResponse response = new CplexResponse(activationMatrix, netBenefit, utility, penalty);
 
 		return response;
@@ -384,7 +393,7 @@ public class Controller {
 			for (ServiceRequest nextRequest : listOfRequestedServices) {
 				s = nextRequest.getServiceID();
 				for (int v = 0; v < vm_types_number; v++) {
-					vm_requests=nextRequest.getVms_requested()[v];
+					vm_requests = nextRequest.getVms_requested()[v];
 					vmRequestMatrix[p][v][s] = vm_requests;
 					total_requests[p][v][s] += vm_requests;
 				}
@@ -412,7 +421,7 @@ public class Controller {
 				for (int n = 0; n < hosts_number; n++) {
 					for (int v = 0; v < vm_types_number; v++) {
 						s = request.getServiceID();
-						vms2DeleteMatrix[n][p][v][s] = generic_scheduler.getRunning_allocations()[n][p][v][s];
+						vms2DeleteMatrix[n][p][v][s] = running_allocations[n][p][v][s];
 					}
 				}
 			}
@@ -421,26 +430,20 @@ public class Controller {
 		return vms2DeleteMatrix;
 	}
 
+	private void reduceRunningAllocation(int[][][][] vmsMatrix) {
 
-	private void reduceRunningAllocation(int[][][][] vmsMatrix ){
-
-		int index=0;
+		int index = 0;
 		for (int n = 0; n < hosts_number; n++)
 			for (int p = 0; p < providers_number; p++)
 				for (int v = 0; v < vm_types_number; v++)
 					for (int s = 0; s < services_number; s++) {
-						index=0;
-						while(index<vmsMatrix[n][p][v][s]){
-							generic_scheduler.getRunning_allocations()[n][p][v][s]--;
+						index = 0;
+						while (index < vmsMatrix[n][p][v][s]) {
+							running_allocations[n][p][v][s]--;
 							index++;
 						}
 					}
-	}	
-
-
-
-
-
+	}
 
 	class DestroyService implements Runnable {
 
@@ -448,8 +451,8 @@ public class Controller {
 		private Boolean deleted = false;
 		Slot slot;
 
-		DestroyService(Slot slot ) {
-			this.slot=slot;
+		DestroyService(Slot slot) {
+			this.slot = slot;
 		}
 
 		@SuppressWarnings({ "rawtypes", "unchecked" })
@@ -457,14 +460,14 @@ public class Controller {
 
 			try {
 
-				String service_name="";
+				String service_name = "";
 				for (int p = 0; p < providers_number; p++) {
 
 					for (ServiceRequest request : this.slot.getServiceRequests2Remove()[p]) {
-						service_name=Utilities.getServiceName(_config,request.slotStart, p, request.serviceID);
+						service_name = Utilities.getServiceName(_config, request.slotStart, p, request.serviceID);
 
-						Hashtable parameters=new Hashtable();
-						parameters.put("service_name",service_name);
+						Hashtable parameters = new Hashtable();
+						parameters.put("service_name", service_name);
 
 						deleted = _webUtilities.destroyService(parameters);
 						Thread.sleep(0);
@@ -493,22 +496,21 @@ public class Controller {
 		int slot;
 		Configuration config;
 
-		LoadService(Configuration config,int slot,int[][][][] activation_matrix) {
+		LoadService(Configuration config, int slot, int[][][][] activation_matrix) {
 
-			this.activation_matrix=activation_matrix;
-			this.slot=slot;
-			this.config=config;
-
+			this.activation_matrix = activation_matrix;
+			this.slot = slot;
+			this.config = config;
 
 		}
 
 		@SuppressWarnings({ "rawtypes", "unchecked" })
 		public void run() {
-			String vm_type_name="";
-			String vm_name="";
-			String vm_series="";
+			String vm_type_name = "";
+			String vm_name = "";
+			String vm_series = "";
 			String service_name;
-			String charm_name="";
+			String charm_name = "";
 			int vms_number;
 			List<String> vm_names_list;
 
@@ -522,26 +524,25 @@ public class Controller {
 						for (int v = 0; v < vm_types_number; v++) {
 							for (int s = 0; s < services_number; s++) {
 
-								service_name=Utilities.buildServiceName(config,slot, p, s);
-								charm_name=config.getService_charm()[s];
-								vm_type_name=Utilities.getVMTypeName(v);
-								vm_name= slot+"_h"+n+"_p"+p+"_v"+v+"_s"+s+"_";
-								vm_series=config.getVm_series();
-								thread_name = "thread_"+vm_name;
+								service_name = Utilities.buildServiceName(config, slot, p, s);
+								charm_name = config.getService_charm()[s];
+								vm_type_name = Utilities.getVMTypeName(v);
+								vm_name = slot + "_h" + n + "_p" + p + "_v" + v + "_s" + s + "_";
+								vm_series = config.getVm_series();
+								thread_name = "thread_" + vm_name;
 
 								vms_number = this.activation_matrix[n][p][v][s];
-								vm_names_list=new ArrayList<String>();
-
+								vm_names_list = new ArrayList<String>();
 
 								for (int i = 0; i < vms_number; i++) {
 
-									vm_name+=i;
+									vm_name += i;
 									vm_names_list.add(vm_name);
 
-									parameters=new Hashtable();
-									parameters.put("vm_name",vm_name);
-									parameters.put("vm_series",vm_series);
-									parameters.put("vm_type",vm_type_name);
+									parameters = new Hashtable();
+									parameters.put("vm_name", vm_name);
+									parameters.put("vm_series", vm_series);
+									parameters.put("vm_type", vm_type_name);
 
 									_webUtilities.createVM(parameters);
 
@@ -549,24 +550,23 @@ public class Controller {
 
 								// Deploy Service in VM 0
 								for (int i = 0; i < vm_names_list.size(); i++) {
-									System.out.println("vm_names_list size"+vm_names_list.size());
-									parameters=new Hashtable();
-									parameters.put("service_name",service_name);
-									parameters.put("vm_name",vm_names_list.get(i));
-									parameters.put("charm_name",charm_name);
+									System.out.println("vm_names_list size" + vm_names_list.size());
+									parameters = new Hashtable();
+									parameters.put("service_name", service_name);
+									parameters.put("vm_name", vm_names_list.get(i));
+									parameters.put("charm_name", charm_name);
 
-									if(i==0) // Deploy in the first VM
+									if (i == 0) // Deploy in the first VM
 										_webUtilities.deployService(parameters);
 									else // Add units for the rest
 										_webUtilities.scaleService(parameters);
 								}
-								
+
 								Thread.sleep(0);
 							}
 						}
 					}
 				}
-
 
 			} catch (Exception e) {
 				System.out.println("Thread " + thread_name + " interrupted.");
@@ -575,10 +575,18 @@ public class Controller {
 			return;
 		}
 
-
 		public SchedulerData getCplexData() {
 			return _cplexData;
 		}
 
 	}
+
+	public int[][][][] getRunning_allocations() {
+		return running_allocations;
+	}
+
+	public int getSlot() {
+		return slot;
+	}
+
 }
